Switch to a faster List Filter arrangement
This makes the workflow slightly more complicated to build, but also
prevents the problem where you can type too fast and hit enter on the
wrong result.

Given that it's effectively a static list, best to let Alfred read it
once and search it itself, rather than re-invoking the Ruby script and
doing some kind of IPC every time a character is typed to update the
list of possible results.

https://www.alfredapp.com/help/workflows/inputs/list-filter/